+++ docs/changes.rst
@@ -12,10 +12,13 @@ Wand Changelog
 Version 0.5.5
 -------------
 
+
 Unreleased.
 
  - Rewrote :meth:`Image.contrast_stretch() <wand.image.BaseImage.contrast_stretch>`
    method to follow modern CLI behavior.
+ - Added :meth:`Image.forward_fourier_transform() <wand.image.BaseImage.forward_fourier_transform>` method.
+ - Added :meth:`Image.inverse_fourier_transform() <wand.image.BaseImage.inverse_fourier_transform>` method.
  - Added ``channel`` parameter support for the following methods.
 
    - :meth:`Image.adaptive_blur() <wand.image.BaseImage.adaptive_blur>`